test: split browser session restore coverage
