+++ cmuxTests/BrowserPanelSessionRestoreTests.swift
@@ -1,0 +1,42 @@
+import Foundation
+import Testing
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+#elseif canImport(cmux)
+@testable import cmux
+#endif
+
+@MainActor
+@Suite(.serialized)
+struct BrowserPanelSessionRestoreTests {
+    @Test
+    func sessionRestoreDefersWebKitLoadUntilPanelIsVisible() throws {
+        let url = try #require(URL(string: "https://example.com/restored"))
+        let panel = BrowserPanel(workspaceId: UUID())
+        defer { panel.close() }
+
+        let originalWebView = panel.webView
+        panel.restoreSessionSnapshot(SessionBrowserPanelSnapshot(
+            urlString: url.absoluteString,
+            profileID: nil,
+            shouldRenderWebView: true,
+            pageZoom: 1.0,
+            developerToolsVisible: false,
+            backHistoryURLStrings: ["https://example.com/back"],
+            forwardHistoryURLStrings: ["https://example.com/forward"]
+        ))
+
+        #expect(panel.webView === originalWebView)
+        #expect(panel.currentURL == url)
+        #expect(!panel.shouldRenderWebView)
+        #expect(panel.webViewLifecycleState == .discarded)
+        #expect(panel.shouldRenderWebViewForSessionSnapshot())
+
+        panel.noteWebViewVisibility(true, reason: "test.visible")
+
+        #expect(panel.shouldRenderWebView)
+        #expect(panel.webViewLifecycleState == .liveVisible)
+        #expect(panel.currentURL == url)
+    }
+}
